Added laradock in order to work with Docker containers. Added the api versioning concept.
