+++ .env.example
@@ -2,12 +2,12 @@ APP_ENV=local
 APP_KEY=
 APP_DEBUG=true
 APP_LOG_LEVEL=debug
-APP_URL=http://api-example.app
+APP_URL=http://192.168.150.100
 
 DB_CONNECTION=mysql
-DB_HOST=127.0.0.1
-DB_PORT=33060
-DB_DATABASE=api_example
+DB_HOST=192.168.150.100
+DB_PORT=3306
+DB_DATABASE=homestead
 DB_USERNAME=homestead
 DB_PASSWORD=secret
