+++ routes/api.php
@@ -14,26 +14,29 @@ use Illuminate\Http\Request;
 */
 
 Route::group(['namespace' => 'Api'], function () {
-    /*
-    |--------------------------------------------------------------------------
-    | Exclusive routes for authentication service
-    |--------------------------------------------------------------------------
-    */
-    Route::group(['namespace' => 'Authentication'], function () {
-        Route::post('/authentication', 'AuthenticationController@authenticate');
-        Route::post('/authentication/forgot_password', 'AuthenticationController@forgotPassword');
-
+    
+    Route::group(['namespace' => 'V1', 'prefix' => 'v1'], function () {
         /*
         |--------------------------------------------------------------------------
-        | Routes with required login
+        | Exclusive routes for authentication service
         |--------------------------------------------------------------------------
         */
-        Route::group(['middleware' => ['jwt.auth']], function () {
+        Route::group(['namespace' => 'Authentication'], function () {
+            Route::post('/authentication', 'AuthenticationController@authenticate');
+            Route::post('/authentication/forgot_password', 'AuthenticationController@forgotPassword');
 
-            Route::put('/authentication/change_password', 'AuthenticationController@changePassword');
-            Route::post('/authentication/logout', 'AuthenticationController@logout');
+            /*
+            |--------------------------------------------------------------------------
+            | Routes with required login
+            |--------------------------------------------------------------------------
+            */
+            Route::group(['middleware' => ['jwt.auth']], function () {
 
-            Route::get('/authentication/refresh_token', 'AuthenticationController@refreshToken');
+                Route::put('/authentication/change_password', 'AuthenticationController@changePassword');
+                Route::post('/authentication/logout', 'AuthenticationController@logout');
+
+                Route::get('/authentication/refresh_token', 'AuthenticationController@refreshToken');
+            });
         });
     });
 });
